# docs/swagger.yaml
basePath: /
definitions:
...
  models.Page:
    type: object
  models.Plugin:
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is the API documentation for the admin panel project.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Admin Panel API
  version: "1.0"
paths:
...
  /pages:
    get:
      description: Retrieve all pages with their details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Page'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all pages
      tags:
      - Pages
    post:
      consumes:
      - application/json
      description: Add a new page with its details
      parameters:
      - description: Page details
        in: body
        name: page
        required: true
        schema:
          $ref: '#/definitions/models.Page'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new page
      tags:
      - Pages
  /pages/{id}:
    delete:
      description: Remove a page by its unique identifier
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a page
      tags:
      - Pages
    get:
      description: Retrieve a single page by its unique identifier
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a page by ID
      tags:
      - Pages
    put:
      consumes:
      - application/json
      description: Update a page's details
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated page details
        in: body
        name: page
        required: true
        schema:
          $ref: '#/definitions/models.Page'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a page
      tags:
      - Pages
  /plugins:
    get:
      description: Retrieve all plugins available in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of plugins
          schema:
            items:
              $ref: '#/definitions/models.Plugin'
            type: array
        "500":
          description: Failed to retrieve plugins
          schema:
            additionalProperties: true
            type: object
      summary: Get all plugins
      tags:
      - Plugins
    post:
      consumes:
      - application/json
      description: Add a new plugin to the system
      parameters:
      - description: Plugin details
        in: body
        name: plugin
        required: true
        schema:
          $ref: '#/definitions/models.Plugin'
      produces:
      - application/json
      responses:
        "201":
          description: Plugin created successfully
          schema:
            $ref: '#/definitions/models.Plugin'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create plugin
          schema:
            additionalProperties: true
            type: object
      summary: Create a new plugin
      tags:
      - Plugins
  /plugins/{id}:
    delete:
      description: Remove a plugin by its unique identifier
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid plugin ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Plugin not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete plugin
          schema:
            additionalProperties: true
            type: object
      summary: Delete a plugin
      tags:
      - Plugins
    put:
      consumes:
      - application/json
      description: Update plugin details by its ID
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated plugin details
        in: body
        name: update
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Plugin updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid plugin ID or request payload
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Plugin not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update plugin
          schema:
            additionalProperties: true
            type: object
      summary: Update a plugin
      tags:
      - Plugins
  /plugins/{id}/enable:
    patch:
      description: Enable a specific plugin by its ID
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Plugin enabled successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid plugin ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Plugin not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to enable plugin
          schema:
            additionalProperties: true
            type: object
      summary: Enable a plugin
      tags:
      - Plugins
  /plugins/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a plugin file (.so) to the system
      parameters:
      - description: Plugin file (.so format)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Plugin uploaded successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid file format or no file uploaded
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to upload plugin file
          schema:
            additionalProperties: true
            type: object
      summary: Upload a new plugin
      tags:
      - Plugins
  /plugins/uploaded:
    get:
      description: Retrieve a list of all uploaded plugin files in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of uploaded plugins
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Failed to retrieve uploaded plugins
          schema:
            additionalProperties: true
            type: object
      summary: List uploaded plugins
      tags:
      - Plugins
...
security:
- BearerAuth: []
swagger: "2.0"